prov/rxd: Warning fixes for 32-bit build
Fixes #2350

+++ prov/rxd/src/rxd_attr.c
@@ -53,7 +53,7 @@ struct fi_ep_attr rxd_ep_attr = {
 	.type = FI_EP_RDM,
 	.protocol = FI_PROTO_RXD,
 	.protocol_version = 1,
-	.max_msg_size = UINT64_MAX,
+	.max_msg_size = SIZE_MAX,
 	.tx_ctx_cnt = 1,
 	.rx_ctx_cnt = 1
 };
